refactor: RoomWebSocketService -> RoomWebSocketHandler 클래스명 변경
